llvm: Enable RTTI support in preparation for new mesa
The mesa 19.0.0 release requires RTTI to allow the build of Nouveau
driver, which is enabled by default when building for x86 and x86-64
target.

This commit enables RTTI support so it prepares the toolchain for the
upcoming change.

(From OE-Core rev: 998e47fc9d04dc4dded4bac5624cdfeb0280c9ac)

Signed-off-by: Otavio Salvador <otavio@ossystems.com.br>
Signed-off-by: Richard Purdie <richard.purdie@linuxfoundation.org>