Revert "gtk+3: add librsvg to DEPENDS"
This was added so that adwaita-icon-theme can find the SVG loader for icon
generation, but the fix is in the wrong place.  GTK+ does not depend on librsvg
at all, the tool in GTK+ which loads the icons uses gdk-pixbuf which uses
loadable modules.

This reverts commit 66828ff04d107b7719c9c8857d7c6c2ebf20a8bb.

(From OE-Core rev: 57c112338068b6e8582480b50b025590fef51052)

Signed-off-by: Ross Burton <ross.burton@intel.com>
Signed-off-by: Richard Purdie <richard.purdie@linuxfoundation.org>